requirements.txt
asgiref==3.7.2
Django==5.0.2
django-autoslug==1.9.9
django-jazzmin==2.6.0
pillow==10.2.0
sqlparse==0.4.4
tzdata==2024.1

gunicorn==20.1.0
whitenoise==6.6.0